requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aioreloader==0.2.2
async-timeout==3.0.1
asyncpg==0.19.0
asyncpgsa==0.26.1
attrs==19.3.0
chardet==3.0.4
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.5.2
PyYAML==5.1.2
SQLAlchemy==1.3.10
url-regex==1.0.0
yarl==1.3.0
uvloop
